# Reduce overhead of idle on SMP