--- a/arch/mips/include/asm/mach-jz4740/jz4780-smp.h
+++ b/arch/mips/include/asm/mach-jz4740/jz4780-smp.h
@@ -86,9 +86,10 @@
 
 #ifdef CONFIG_SMP
 
+extern void jz4780_smp_wait_irqoff(void);
+
 extern void jz4780_smp_init(void);
 extern void jz4780_secondary_cpu_entry(void);
-extern void jz4780_smp_irq(void);
 
 #else /* !CONFIG_SMP */